+++ tests/test-current.js
@@ -1,6 +1,7 @@
 // const Promise = require('bluebird')
 // const enums = require('./enums.spec')
 // const is = require('./is.spec')
+const datetime = require('./datetime.spec')
 // const dbAssertDatabase = require('./db-assert-database.spec')
 // const dbAssertTable = require('./db-assert-table.spec')
 // const dbAssertIndex = require('./db-assert-index.spec')
@@ -20,7 +21,7 @@ const dbAssert = require('./db-assert.spec')
 // const queueAddJob = require('./queue-add-job.spec')
 // const queueGetJob = require('./queue-get-job.spec')
 // const queueGetNextJob = require('./queue-get-next-job.spec')
-const queueProcess = require('./queue-process.spec')
+// const queueProcess = require('./queue-process.spec')
 // const queueChange = require('./queue-change.spec')
 // const queueInterruption = require('./queue-interruption.spec')
 // const queueCancelJob = require('./queue-cancel-job.spec')
@@ -32,5 +33,5 @@ const queueProcess = require('./queue-process.spec')
 
 return dbAssert().then(() => {
 }).then(() => {
-  return queueProcess()
+  return datetime()
 })