+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *   Added 'db-driver' module for connection testing.
 *   Changed 'Queue()' constructor API to add connection options.
 *   Standardized time related options to milliseconds.
+*   Added 'datetime' module to add and format dates.
 
 ## v0.0.8 / 2016-09-09